+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/shelldescriptor/application/DecentralRegistryService.java
@@ -20,4 +20,6 @@ package org.eclipse.tractusx.traceability.shelldescriptor.application;
 
 public interface DecentralRegistryService {
     void updateShellDescriptorAndSynchronizeAssets();
+
+    void deleteAll();
 }