+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/shelldescriptor/application/ShellDescriptorService.java
@@ -24,4 +24,6 @@ import java.util.List;
 
 public interface ShellDescriptorService {
     List<ShellDescriptor> determineExistingShellDescriptorsAndUpdate(List<ShellDescriptor> ownShellDescriptors);
+
+    void deleteAll();
 }