+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Refactored asset structure to allow new API for assets-as-planned
 - Optional parameter "receiverBpn" to /investigations endpoint
 - NEW API /api/assets/as-planned
+- NEW API DELETE /api/registry
 
 ### Changed
 - API BREAKING CHANGE: /api/assets changed to /api/assets/as-built